tests: Move custom frontend injection to testenv
This opens us up to support other means of managing custom frontends
than just buildkit v0.12 + inputs.
It doesn't add anything new, but makes what we have more open to it.
